refactor: remove finale tab and game form dialog components
- Deleted the finale tab component and its associated files including HTML, CSS, and spec files.
- Removed the finale game form dialog component and its related files.
- Added a new finale game form dialog component under the poules tab with updated HTML and TypeScript files.
- Updated poules tab to include finale functionality with new UI elements for managing finale games.
- Introduced new methods for handling finale game logic in the poules tab component.
- Adjusted styles for the finale bracket and match display in the poules tab.
