+++ src/app/admin/types/shared/settings/settings.html
@@ -1,5 +1,5 @@
 @if (role() === 'admin') {
-  <div class="flex flex-col gap-8">
+  <div class="flex flex-col gap-8 my-3">
     <section>
       <h2 class="text-xl font-semibold mb-4! mx-3!">
         {{ 'admin.general.title' | transloco }}
