+++ src/app/shared/tournament-tabs/tournament-tabs.css
@@ -1,4 +1,3 @@
 :host {
-  margin-bottom: 1em;
   display: block;
 }
