+++ src/app/tournaments/types/shared/dashboard/tournament-dashboard.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   gap: 1rem;
-  padding: 0 1rem 1rem;
+  padding: 1rem;
   max-width: 1200px;
   margin: 0 auto;
 }
