+++ src/app/tournaments/types/shared/games/games.html
@@ -1,7 +1,7 @@
 <p-confirmdialog></p-confirmdialog>
 
 @if (hasPoules()) {
-  <div class="flex flex-wrap items-center gap-2 mx-3 mb-3">
+  <div class="flex flex-wrap items-center gap-2 m-3">
     <p-select
       class="w-[calc(50%-0.25rem)] sm:w-auto min-w-0 sm:min-w-40 flex-none"
       [options]="sortedTeams()"
