+++ src/app/tournaments/types/shared/poules-tab/poules-tab.ts
@@ -1,8 +1,9 @@
-import { Component, computed, inject, linkedSignal } from '@angular/core';
+import { Component, computed, inject, input, linkedSignal } from '@angular/core';
 import { AccordionModule } from 'primeng/accordion';
 import { CardModule } from 'primeng/card';
+import { TabsModule } from 'primeng/tabs';
 import { Team } from '../teams/teams';
-import { Poule, Serie } from '../../poules/poules';
+import { Poule, Serie, FinaleGame } from '../../poules/poules';
 import { NgTemplateOutlet } from '@angular/common';
 import { ApplyPipe } from 'ngxtension/call-apply';
 import { DocumentReference } from '@angular/fire/firestore';
@@ -13,9 +14,17 @@ import { ConfirmationService } from 'primeng/api';
 import { ConfirmDialogModule } from 'primeng/confirmdialog';
 import { DialogService } from 'primeng/dynamicdialog';
 import { TooltipModule } from 'primeng/tooltip';
+import { Select } from 'primeng/select';
+import { FormsModule } from '@angular/forms';
 import { SerieFormDialog } from './serie-form-dialog/serie-form-dialog';
 import { PouleFormDialog } from './poule-form-dialog/poule-form-dialog';
 import { TeamPouleDialog } from './team-poule-dialog/team-poule-dialog';
+import {
+  FinaleGameFormDialog,
+  SaveFinaleGameEvent,
+} from './finale-game-form-dialog/finale-game-form-dialog';
+
+export type { SaveFinaleGameEvent };
 import { PoulesStore } from '../../../../store/poules.store';
 import { AuthStore } from '../../../../store/auth.store';
 import { TournamentActionsService } from '../../../../shared/services/tournament-actions.service';
@@ -41,6 +50,21 @@ export interface TeamInPouleEvent {
   teamRef: DocumentReference;
 }
 
+export interface SetFinaleSizeEvent {
+  serieRef: DocumentReference;
+  size: number;
+}
+
+export interface GenerateFinaleEvent {
+  serieRef: DocumentReference;
+  serieName: string;
+  finaleSize: number;
+}
+
+export interface DeleteFinaleGamesEvent {
+  serieRef: DocumentReference;
+}
+
 export interface TeamStanding {
   ref: DocumentReference;
   name: string;
@@ -50,9 +74,16 @@ export interface TeamStanding {
   pointsConceded: number;
 }
 
+export interface BracketRound {
+  round: string;
+  roundOrder: number;
+  games: FinaleGame[];
+}
+
 @Component({
   selector: 'app-poules-tab',
   imports: [
+    TabsModule,
     AccordionModule,
     CardModule,
     NgTemplateOutlet,
@@ -62,6 +93,8 @@ export interface TeamStanding {
     Message,
     ConfirmDialogModule,
     TooltipModule,
+    FormsModule,
+    Select,
   ],
   providers: [DialogService, ConfirmationService],
   templateUrl: './poules-tab.html',
@@ -78,6 +111,7 @@ export class PoulesTab {
   teams = this.poulesStore.teams;
   series = this.poulesStore.series;
   role = this.authStore.role;
+  showFinale = input(false);
 
   sortedSeries = computed(() =>
     [...this.series()]
@@ -93,10 +127,17 @@ export class PoulesTab {
       })),
   );
 
-  openSeriesIds = linkedSignal(() => {
+  activeSerie = linkedSignal(() => {
     const series = this.sortedSeries();
-    return series.length === 1 ? [series[0].ref.id] : [];
+    return series.length > 0 ? series[0].ref.id : '';
   });
+
+  sizeOptions = [2, 4, 8, 16, 32].map((v) => ({ label: String(v), value: v }));
+
+  onSerieTabChange(event: any): void {
+    // PrimeNG tabs passes the value directly in the event
+    this.activeSerie.set(event);
+  }
 
   private computeStandings(poule: Poule): TeamStanding[] {
     const teams = this.teams();
@@ -374,6 +415,109 @@ export class PoulesTab {
       accept: () => {
         void this.tournamentActions.removeTeamFromPoule({ poule, teamRef });
       },
+    });
+  }
+
+  // Finale methods
+  getBracketRounds = (serie: Serie): BracketRound[] => {
+    const games = serie.finaleGames ?? [];
+    const roundMap = new Map<string, BracketRound>();
+    for (const game of games) {
+      if (!roundMap.has(game.round)) {
+        roundMap.set(game.round, { round: game.round, roundOrder: game.roundOrder, games: [] });
+      }
+      roundMap.get(game.round)!.games.push(game);
+    }
+    return [...roundMap.values()]
+      .sort((a, b) => b.roundOrder - a.roundOrder)
+      .map((r) => ({ ...r, games: r.games.sort((a, b) => a.matchNumber - b.matchNumber) }));
+  };
+
+  getTeamNameForFinale = (
+    ref: DocumentReference | null | undefined,
+    placeholder?: string,
+  ): string => {
+    if (ref) {
+      const team = this.teams().find((t) => t.ref.id === ref.id);
+      if (team) return team.name;
+    }
+    if (placeholder) {
+      if (placeholder.startsWith('finale.winnerOf:')) {
+        const parts = placeholder.split(':');
+        const roundKey = parts[1];
+        const matchNum = parts[2];
+        const roundLabel = this.translocoService.translate(roundKey);
+        return this.translocoService.translate('finale.winnerOf', {
+          round: roundLabel,
+          match: matchNum,
+        });
+      }
+      return placeholder;
+    }
+    return this.translocoService.translate('finale.noTeam');
+  };
+
+  onSetFinaleSize(serie: Serie, size: number | null): void {
+    if (size == null) return;
+    void this.tournamentActions.setFinaleSize({ serieRef: serie.ref, size });
+  }
+
+  onGenerateFinale(serie: Serie): void {
+    if (!serie.finaleSize) return;
+    if (serie.finaleGames && serie.finaleGames.length > 0) {
+      this.confirmationService.confirm({
+        message: this.translocoService.translate('finale.regenerateConfirm'),
+        header: this.translocoService.translate('shared.confirm.deleteHeader'),
+        icon: 'pi pi-exclamation-triangle',
+        accept: () => {
+          void this.tournamentActions.generateFinale({
+            serieRef: serie.ref,
+            serieName: serie.name,
+            finaleSize: serie.finaleSize!,
+          });
+        },
+      });
+    } else {
+      void this.tournamentActions.generateFinale({
+        serieRef: serie.ref,
+        serieName: serie.name,
+        finaleSize: serie.finaleSize,
+      });
+    }
+  }
+
+  onDeleteFinaleGames(serie: Serie): void {
+    this.confirmationService.confirm({
+      message: this.translocoService.translate('finale.deleteConfirm'),
+      header: this.translocoService.translate('shared.confirm.deleteHeader'),
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        void this.tournamentActions.deleteFinaleGames({ serieRef: serie.ref });
+      },
+    });
+  }
+
+  onEditFinaleGame(game: FinaleGame): void {
+    const ref = this.dialogService.open(FinaleGameFormDialog, {
+      header: this.translocoService.translate('finale.dialogEditGame'),
+      data: {
+        teams: this.teams(),
+        role: this.role(),
+        isEditing: true,
+        gameRef: game.ref,
+        initialTeam1Ref: game.refTeam1 ?? null,
+        initialTeam2Ref: game.refTeam2 ?? null,
+        initialScoreTeam1: game.scoreTeam1 ?? null,
+        initialScoreTeam2: game.scoreTeam2 ?? null,
+        team1Placeholder: game.team1Placeholder,
+        team2Placeholder: game.team2Placeholder,
+      },
+      width: '500px',
+    });
+    ref?.onClose.subscribe((result: SaveFinaleGameEvent | undefined) => {
+      if (result) {
+        void this.tournamentActions.saveFinaleGame(result);
+      }
     });
   }
 }
