+++ src/app/tournaments/types/shared/teams/teams.html
@@ -1,7 +1,7 @@
 <p-confirmdialog></p-confirmdialog>
 
 @if (teams().length > 0) {
-  <div class="flex flex-wrap items-center gap-2 mx-3 mb-3">
+  <div class="flex flex-wrap items-center gap-2 m-3">
     <p-iconfield>
       <p-inputicon class="pi pi-search" />
       <input
